fix(radar): resolve tensor size and data type mismatches
- Fix embed_dim mismatch by setting radar encoder embed_dim to 256 to match transformer_width
- Convert radar data to float32 to avoid double/float type mismatch
- Update context_length to 248 to match radar encoder sequence length (31*8=248)
- Ensure positional embedding size matches radar encoder output sequence length

+++ src/datasets/base_dataset.py
@@ -124,10 +124,10 @@ def load_radar_data(npz_file_path, opt):
 
         # Return as dictionary with three separate views
         return {
-            'range_time': range_processed,    # (256, T_processed)
-            'doppler_time': doppler_processed,  # (128, T_processed)
-            'azimuth_time': azimuth_processed,  # (128, T_processed)
-            'mask': range_mask,  # All views should have same T after processing
+            'range_time': range_processed.astype(np.float32),    # (256, T_processed)
+            'doppler_time': doppler_processed.astype(np.float32),  # (128, T_processed)
+            'azimuth_time': azimuth_processed.astype(np.float32),  # (128, T_processed)
+            'mask': range_mask.astype(np.float32),  # All views should have same T after processing
             'T': range_T
         }
